Fix test module printf warning

+++ src/test.c
@@ -8,8 +8,7 @@ static void test_run_individual(test_t *t, test_function_t test_function) {
     if (!t->failed) {
         printf("\x1b[1;32m[PASSED]\x1b[0m %s\n", test_function.name);
         t->num_successes++;
-    }
-    else {
+    } else {
         printf("\x1b[1;31m[FAILED]\x1b[0m %s\n", test_function.name);
     }
 }
@@ -30,7 +29,7 @@ int test_run(test_function_t test_functions[], size_t num_tests) {
     if (test.failed) {
         puts("\x1b[31m1 failed.\x1b[0m");
         printf("\x1b[33m%d left to be tested.\x1b[0m\n",
-               num_tests - test.num_successes - 1);
+               (int) num_tests - test.num_successes - 1);
     }
 
     return test.failed;
